Remove chroots after build is done (even on failure)
When a build starts, the previous chroot is deleted. There is no reason
to keep the chroot around, and it just takes up disk space on the ARM
slaves. Even if the build fails, we don't typically use the build
slaves to debug such problems.

This shouldn't affect tmpfs builds at all